% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.1row.R
\name{estimate.1row}
\alias{estimate.1row}
\title{Impute one row of data}
\usage{
estimate.1row(row, s, m)
}
\arguments{
\item{row}{A vector}

\item{s}{Covariance matrix}

\item{m}{Mean vector}
}
\value{
A list with the estimated mean and variance for the data
}
\description{
Impute one row of data
}
