% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.norm2.R
\name{prelim.norm2}
\alias{prelim.norm2}
\title{Fixed version of prelim.norm from package norm}
\usage{
prelim.norm2(x)
}
\arguments{
\item{x}{A data matrix.}
}
\value{
cf norm::prelim.norm
}
\description{
This function fixes the way that mdp is generated so that x can have more than 32 columns
}
