% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMVN_impute.R
\name{SIMVN_impute}
\alias{SIMVN_impute}
\title{Impute missing data by the conditional mean of the observed values, using a normal approximation}
\usage{
SIMVN_impute(fittedM, X)
}
\arguments{
\item{fittedM}{A parameter list as returned by SIMVN_train}

\item{X}{A data matrix}
}
\value{
\code{X} filled with estimated values
}
\description{
Impute missing data by the conditional mean of the observed values, using a normal approximation
}
