% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMVN_train.R
\name{SIMVN_train}
\alias{SIMVN_train}
\title{Estimate data mean and variance for imputation}
\usage{
SIMVN_train(X, verbose = F)
}
\arguments{
\item{X}{A data matrix.}

\item{verbose}{A boolean: whether or not to print EM iterations}
}
\value{
A list with the estimated mean and variance for the data
}
\description{
norm::rngseed MUST be run before the function is used, otherwise it will not work
}
