% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_matrix.R
\name{to_matrix}
\alias{to_matrix}
\title{Converts a dataframe or vector to matrix of the right orientation, handling the case where the data is 1D}
\usage{
to_matrix(x, horiz)
}
\arguments{
\item{x}{a dataframe or vector}

\item{horiz}{Wheter the returned matrix should be a row matrix}
}
\value{
\code{X} filled with estimated values
}
\description{
Converts a dataframe or vector to matrix of the right orientation, handling the case where the data is 1D
}
